+++ bean-searcher-doc/docs/.vitepress/config/zh.ts
@@ -13,7 +13,7 @@ export default defineConfig({
     nav: [
       { text: '指南', link: '/guide/info/bean-searcher', activeMatch: '/guide/' },
       {
-        text: 'v4.5.0',
+        text: 'v4.5.2',
         items: [
           {
             text: '更新日志',
@@ -22,6 +22,21 @@ export default defineConfig({
           {
             text: '参与贡献',
             link: 'https://gitee.com/troyzhxu/bean-searcher/blob/dev/CONTRIBUTING.md'
+          }
+        ]
+      },
+      {
+        text: '生态',
+        items: [
+          {
+            text: 'Bean Searcher Label',
+            link: '/zoo/label/info',
+            activeMatch: '/zoo/'
+          },
+          {
+            text: 'Bean Searcher Exporter',
+            link: '/zoo/exporter/info',
+            activeMatch: '/zoo/'
           }
         ]
       },
@@ -115,7 +130,32 @@ export default defineConfig({
             ]
           }
         ]
-      }
+      },
+      '/zoo/': {
+        base: '/zoo', 
+        items: [
+          {
+            text: 'Bean Searcher Label',
+            base: '/zoo/label',
+            collapsed: false,
+            items: [
+              { text: '介绍', link: '/info' },
+              { text: 'Bean Searcher', link: '/bean-searcher' },
+              { text: 'What\'s New?', link: '/versions' },
+            ]
+          },
+          {
+            text: 'Bean Searcher Exporter',
+            base: '/zoo/exporter',
+            collapsed: false,
+            items: [
+              { text: '介绍', link: '/info' },
+              { text: '集成', link: '/integration' },
+              { text: '使用', link: '/use' },
+            ]
+          },
+        ]
+      },
     },
 
     outline: {